+++ CdIts.Caffoa.Cli/Model/SchemaItem.cs
@@ -5,7 +5,8 @@ public class SchemaItem
     public enum ObjectType
     {
         Regular,
-        StringEnum
+        StringEnum,
+        Array
     };
     public SchemaItem(string name, string className)
     {