+++ changelog.md
@@ -2,6 +2,8 @@
 ## 4.9.0
 * extracted AzFuncTestHelper to a separate repository
 * Add CaffoaClientStringException and CAffoaClientJsonException\<T> for easier error error handling in servers
+* allow to create objects with anyOf
+* allow to create array objects
 
 ## 4.8.1
 * fix an error with enum lists that references enums from another yaml file